move implementation to cpp files + div.
- moved network messages implementation into cpp files.
- udp client (unfinished).
- typedef network_to_native and visa versa functions.
