+++ include/discnet_lib/network/messages/data_message.hpp
@@ -15,12 +15,6 @@ namespace discnet::network::messages
     {
         uint16_t m_identifier = init_required;
         std::vector<discnet::byte_t> m_buffer = {};
-
-        bool operator==(const data_message_t& message) const
-        {
-            return m_identifier == message.m_identifier &&
-                std::equal(m_buffer.begin(), m_buffer.end(), message.m_buffer.begin(), message.m_buffer.end());
-        }
     };
 
     using expected_data_message_t = std::expected<data_message_t, std::string>;
@@ -30,45 +24,10 @@ namespace discnet::network::messages
         static const size_t s_data_message_header_size = 4;
         static const message_type_e s_message_type = message_type_e::data_message;
 
-        static size_t encoded_size(const data_message_t& message)
-        {
-            const size_t s_header_size = header_codec_t::size();
-            const size_t message_size = s_data_message_header_size + message.m_buffer.size();
-            return s_header_size + message_size;
-        }
-
-        static bool encode(network::buffer_t& buffer, const data_message_t& message)
-        {
-            const size_t message_size = encoded_size(message);
-            if (buffer.remaining_bytes() < message_size)
-            {
-                return false;
-            }
-
-            header_codec_t::encode(buffer, message_size, s_message_type);
-
-            buffer.append(boost::endian::native_to_big((uint16_t)message.m_identifier));
-            buffer.append(boost::endian::native_to_big((uint16_t)message.m_buffer.size()));
-            buffer.append(std::span(message.m_buffer.data(), message.m_buffer.size()));
-
-            return true;
-        }
-
-        static expected_data_message_t decode(network::buffer_t& buffer)
-        {
-            if (buffer.bytes_left_to_read() < s_data_message_header_size)
-            {
-                return std::unexpected("not enough bytes in buffer to read data_message.");
-            }
-            
-            uint16_t identifier = boost::endian::big_to_native(buffer.read_uint16());
-            uint16_t size = boost::endian::big_to_native(buffer.read_uint16());
-            buffer_span_t data = buffer.read_buffer(size);
-            
-            data_message_t result{.m_identifier = identifier};
-            result.m_buffer.assign(data.begin(), data.end());
-
-            return result;
-        }
+        DISCNET_EXPORT static size_t encoded_size(const data_message_t& message);
+        DISCNET_EXPORT static bool encode(network::buffer_t& buffer, const data_message_t& message);
+        DISCNET_EXPORT static expected_data_message_t decode(network::buffer_t& buffer);
     };
+
+    DISCNET_EXPORT bool operator==(const data_message_t& lhs, const data_message_t& rhs);
 } // !namespace discnet::network::messages
