+++ include/discnet_lib/typedefs.hpp
@@ -5,6 +5,7 @@
 #pragma once
 
 #include <memory>
+#include <boost/endian.hpp>
 #include <boost/asio.hpp>
 #include <boost/uuid/uuid.hpp>
 #include <discnet_lib/discnet_lib.hpp>
@@ -27,4 +28,19 @@ namespace discnet
         template <class T>
         operator T() const { static_assert(typeof(T) == 0, "struct memeber not initialized"); }
     } static const init_required;
-}
+
+    namespace network
+    {
+        template <typename ... args_t>
+        auto native_to_network(args_t&&... args) -> decltype(boost::endian::native_to_big(std::forward<args_t>(args)...))
+        {
+            return boost::endian::native_to_big(std::forward<args_t>(args)...);
+        }
+
+        template <typename ... args_t>
+        auto network_to_native(args_t&&... args) -> decltype(boost::endian::big_to_native(std::forward<args_t>(args)...))
+        {
+            return boost::endian::big_to_native(std::forward<args_t>(args)...);
+        }
+    } // ! namespace network
+} // ! namespace discnet
